% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetCor.R
\name{targetCor}
\alias{targetCor}
\title{Significance-filtered knowledge-based target.}
\usage{
targetCor(x, genegroups)
}
\arguments{
\item{x}{A \eqn{n \times p} data matrix.}

\item{genegroups}{A list of genes obtained using the database KEGG, where
each entry itself is a list of pathway names this genes belongs to. If a
gene does not belong to any gene functional group, the entry is NA.}
}
\value{
A \eqn{p \times p}{p x p} matrix.
}
\description{
The \eqn{p \times p}{p x p} target Cor is computed from the \eqn{n \times
p}{n x p} data matrix. It it a modified version of target G. In particular,
it tests the correlations (with a significance level of 0.05) and sets the
non-significant correlations to zero before the mean correlation
\eqn{\bar{r}}{r} is computed.
}
\details{
For retained linked pairs, the off-diagonal target is
\deqn{T_{ij} = \bar{r}_{sig}\sqrt{s_{ii}s_{jj}}}{T_ij = r_sig * sqrt(s_ii * s_jj)}
where \eqn{\bar{r}_{sig}}{r_sig} is the mean correlation among links
whose correlation test p-value is below 0.05.
}
\examples{

# A short example on a toy dataset
# require(SHIP)
data(expl)
target_matrix <- targetCor(expl$x, expl$genegroups)
which(target_matrix[upper.tri(target_matrix)] != 0) # not many non zero coefficients !
}
\seealso{
Other covariance targets:
\code{\link[=targetD]{targetD()}},
\code{\link[=targetF]{targetF()}},
\code{\link[=targetG]{targetG()}},
\code{\link[=targetGpos]{targetGpos()}},
\code{\link[=targetGstar]{targetGstar()}}
}
\author{
Monika Jelizarow and Vincent Guillemot
}
\concept{covariance targets}
\keyword{methods}
\keyword{multivariate}
