% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetD.R
\name{targetD}
\alias{targetD}
\title{Diagonal covariance target.}
\usage{
targetD(x, genegroups = NULL)
}
\arguments{
\item{x}{A \eqn{n \times p}{n x p} data matrix.}

\item{genegroups}{The genegroups are not used for this target.}
}
\value{
A \eqn{p \times p}{p x p} diagonal matrix.
}
\description{
The target keeps the sample variances and sets all covariances to zero. If
\eqn{S = (s_{ij})}{S = (s_ij)} is the sample covariance matrix, the target
is
\deqn{T_{ij} = \begin{cases} s_{ii} & \text{if } i = j \\ 0 & \text{otherwise} \end{cases}}{T_ij = s_ii if i = j, and 0 otherwise.}
}
\examples{

x <- matrix(rnorm(10*30),10,30)
target_matrix <- targetD(x, NULL)

}
\seealso{
Other covariance targets:
\code{\link[=targetCor]{targetCor()}},
\code{\link[=targetF]{targetF()}},
\code{\link[=targetG]{targetG()}},
\code{\link[=targetGpos]{targetGpos()}},
\code{\link[=targetGstar]{targetGstar()}}
}
\author{
Monika Jelizarow and Vincent Guillemot
}
\concept{covariance targets}
\keyword{methods}
\keyword{multivariate}
