% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetGpos.R
\name{targetGpos}
\alias{targetGpos}
\title{Positive-correlation knowledge-based target.}
\usage{
targetGpos(x, genegroups)
}
\arguments{
\item{x}{A \eqn{n \times p}{n x p} data matrix.}

\item{genegroups}{A list of genes obtained using the database KEGG, where
each entry itself is a list of pathway names this genes belongs to. If a
gene does not belong to any gene functional group, the entry is NA.}
}
\value{
A \eqn{p \times p}{p x p} matrix.
}
\description{
This is a version of \code{\link[=targetG]{targetG()}} that estimates the common correlation from
positive linked correlations only. Negative linked correlations contribute
zero to the off-diagonal target.
\deqn{T_{ij} = \bar{r}_+\sqrt{s_{ii}s_{jj}}}{T_ij = r_plus * sqrt(s_ii * s_jj)}
for linked pairs with positive correlation, where \eqn{\bar{r}_+}{r_plus}
is their mean positive correlation.
}
\examples{
# require(SHIP)
data(expl)
target_matrix <- targetGpos(expl$x, expl$genegroups)
which(target_matrix[upper.tri(target_matrix)] != 0) # not many non zero coefficients !

}
\seealso{
Other covariance targets:
\code{\link[=targetCor]{targetCor()}},
\code{\link[=targetD]{targetD()}},
\code{\link[=targetF]{targetF()}},
\code{\link[=targetG]{targetG()}},
\code{\link[=targetGstar]{targetGstar()}}
}
\author{
Monika Jelizarow and Vincent Guillemot
}
\concept{covariance targets}
\keyword{methods}
\keyword{multivariate}
