% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetGstar.R
\name{targetGstar}
\alias{targetGstar}
\title{Signed-correlation knowledge-based target.}
\usage{
targetGstar(x, genegroups)
}
\arguments{
\item{x}{A \eqn{n \times p} data matrix.}

\item{genegroups}{A list of genes obtained using the database KEGG, where
each entry itself is a list of pathway names this genes belongs to. If a
gene does not belong to any gene functional group, the entry is NA.}
}
\value{
A \eqn{p \times p}{p x p} matrix.
}
\description{
The \eqn{p \times p}{p x p} target Gstar is computed from the \eqn{n \times
p}{n x p} data matrix. It it a modified version of target G. In particular,
it involves two parameters for the correlation (a positive and a negative
one) instead of the single parameter \eqn{\bar{r}}{r} in order to account
for negatively correlated genes within the same pathway.
}
\details{
The target uses separate mean correlations for positive and negative linked
pairs:
\deqn{T_{ij} = \begin{cases}
  \bar{r}_+\sqrt{s_{ii}s_{jj}} & \text{if } r_{ij} > 0 \\
  \bar{r}_-\sqrt{s_{ii}s_{jj}} & \text{if } r_{ij} < 0
\end{cases}}{T_ij = r_plus * sqrt(s_ii * s_jj) if r_ij > 0, and r_minus * sqrt(s_ii * s_jj) if r_ij < 0.}
Here \eqn{\bar{r}_+}{r_plus} and \eqn{\bar{r}_-}{r_minus} are the mean
positive and negative linked correlations.
}
\examples{
# require(SHIP)
data(expl)
target_matrix <- targetGstar(expl$x, expl$genegroups)
which(target_matrix[upper.tri(target_matrix)] != 0) # not many non zero coefficients !

}
\seealso{
Other covariance targets:
\code{\link[=targetCor]{targetCor()}},
\code{\link[=targetD]{targetD()}},
\code{\link[=targetF]{targetF()}},
\code{\link[=targetG]{targetG()}},
\code{\link[=targetGpos]{targetGpos()}}
}
\author{
Monika Jelizarow and Vincent Guillemot
}
\concept{covariance targets}
\keyword{methods}
\keyword{multivariate}
